+++ reader/src/ASTR/menupage/oritem.vue
@@ -6,6 +6,14 @@
             <div class="setnumber">#{{eventData.set}}</div>
         </div>
         <div class="orstories" v-show="showStory">
+            <div class="exportBar nt">
+                <div @click="exportAll()">
+                    <span class="material-icons">
+                        file_download
+                    </span>
+                    <span>{{i18n.export2excel[lang]}}</span>
+                </div>
+            </div>
             <div class="orstory" v-for="(story,sidx) in eventData.infoUnlockDatas" :key="sidx" @click="loadStory(story.storyTxt)">
                 {{sidx}}. {{story.storyName}}
             </div>
@@ -15,10 +23,13 @@
 
 <script>
 import func from "../func"
+import i18n from '../i18n.json';
 export default {
     data(){
         return{
-            server: func.server
+            server: func.server,
+            i18n: i18n,
+            lang: func.l,
         }
     },
     emits: ['focusme'],
@@ -27,6 +38,25 @@ export default {
         showStory: Boolean
     },
     methods:{
+            exportAll(){
+                let data = this.eventData.infoUnlockDatas;
+                let eventname = this.eventData.name;
+                let setnumber = this.eventData.set;
+                window.localStorage.setItem("filename", eventname+'_'+setnumber);
+                let server = this.server;
+                let exportList = [];
+                data.forEach(story => {
+                    exportList.push({
+                        server: server,
+                        path: story.storyTxt,
+                        storyCode: story.storyCode,
+                        avgTag: story.avgTag,
+                        storyName: story.storyName
+                    });
+                });
+                window.localStorage.setItem("exportList", JSON.stringify(exportList));
+                window.location.href = "export.html";
+        },
         getAvatar(){
             return 'https://aceship.github.io/AN-EN-Tags/img/avatars/char_'+this.eventData.cid+'_'+this.eventData.cin+'.png'
         },
@@ -81,6 +111,9 @@ export default {
 }
 .orstory:hover{
     background-color: rgba(255,255,255,0.2);
+}
+.orstory .exportBar{
+    margin: 2px;
 }
 @media(max-width: 1000px){
     .orcard img{