+++ reader/src/export.js
@@ -1,0 +1,5 @@
+import { createApp } from 'vue'
+import exportApp from './export/export.vue'
+
+
+createApp(exportApp).mount('#export')